update husky todolist
