+++ src/pages/Husky/TodoList/Add.jsx
@@ -1,4 +1,4 @@
-import { AddIcon } from './components/icon';
+import { ReactComponent as AddIcon } from './components/icon/AddIcon.svg';
 import { Input, Button } from './components/utils';
 
 export function Add({ setNewTalk, newTalk, methods }) {
@@ -6,12 +6,18 @@ export function Add({ setNewTalk, newTalk, methods }) {
     <div className="flex gap-2">
       <div className="relative">
         <span className="absolute top-4 left-4">
-          <AddIcon />
+          <AddIcon
+            className="h-5 w-5 text-gray-400"
+            stroke="currentColor"
+            strokeWidth="2"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          />
         </span>
         <Input
           value={newTalk}
           placeholder="new task describe"
-          onInput={(e) => setNewTalk(e.target.value)}
+          onInput={(e) => setNewTalk(e.target.value.trim())}
           icon={true}
         />
       </div>
