+++ src/pages/Husky/TodoList/Item.jsx
@@ -1,32 +1,33 @@
-import { DeleteIcon, EditIcon } from './components/icon';
+import { ReactComponent as EditIcon } from './components/icon/EditIcon.svg';
+import { ReactComponent as DeleteIcon } from './components/icon/DeleteIcon.svg';
 import { Input } from './components/utils';
 
-export function Item({ task, onClick, clickDelete, clickEdit, editTask }) {
-  function editToggle(edit) {
-    if (edit) {
-      return <Input value={task.editTitle} onInput={(e) => editTask(e.target.value)} />;
-    }
-    return (
-      <label className="flex flex-1 items-center space-x-3 pl-1">
-        <input
-          type="checkbox"
-          className="bg-check h-6 w-6 rounded-md border border-gray-300 bg-white checked:border-transparent focus:outline-none"
-          onChange={onClick}
-        />
-        <span className={'font-normal text-gray-700' + (task.status ? ' line-through' : '')}>
-          {task.title}
-        </span>
-      </label>
-    );
+function EditToggle({ task, onClick, editTask }) {
+  if (task.edit) {
+    return <Input value={task.editTitle} onInput={(e) => editTask(e.target.value)} />;
   }
+  return (
+    <label className="flex flex-1 items-center space-x-3 pl-1">
+      <input
+        type="checkbox"
+        className="bg-check h-6 w-6 rounded-md border border-gray-300 bg-white checked:border-transparent focus:outline-none"
+        onChange={onClick}
+      />
+      <span className={'font-normal text-gray-700' + (task.status ? ' line-through' : '')}>
+        {task.title}
+      </span>
+    </label>
+  );
+}
 
+export function Item({ task, onClick, clickDelete, clickEdit, editTask }) {
   return (
     <li className="mb-2 flex flex-row border-gray-400">
       <div className="flex flex-1 transform select-none items-center gap-4 rounded-md bg-white p-4 shadow transition duration-500 ease-in-out hover:-translate-y-1 hover:shadow-lg">
-        {editToggle(task.edit)}
+        <EditToggle task={task} onClick={onClick} editTask={editTask} />
         <button className="flex w-12 justify-end gap-2 text-right">
-          {!task.status && <EditIcon onClick={clickEdit} />}
-          <DeleteIcon onClick={clickDelete} />
+          {!task.status && <EditIcon className="h-5 w-5 text-gray-400" onClick={clickEdit} />}
+          <DeleteIcon className="h-5 w-5 text-gray-400" onClick={clickDelete} />
         </button>
       </div>
     </li>
