--- src/pages/Husky/TodoList/components/icon/add.jsx
@@ -1,13 +1,0 @@
-export function AddIcon() {
-  return (
-    <svg className="h-5 w-5 text-gray-400" viewBox="0 0 24 24">
-      <path
-        d="M19 13h-6v6h-2v-6H5v-2h6V5h2v6h6v2z"
-        stroke="currentColor"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  );
-}
