--- src/pages/Husky/TodoList/components/icon/index.js
@@ -1,3 +1,0 @@
-export * from './add';
-export * from './delete';
-export * from './edit';
